fix: GUI blocks forwarding of keys presses
